integrate api in appointment calendar + break down file
